data-export/requirements.txt
requests==2.31.0
minio==7.2.0
psycopg2-binary==2.9.9